# specs/001-screener-mvp/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: US Stock Screener API
  version: 0.1.0
  description: |
    Backend API for the US Stock Screener MVP (spec 001-screener-mvp).

    Conventions:
    - Every response that exposes user-visible data carries `data_as_of` and `disclaimer` fields per Constitution V.
    - All monetary values are USD; serialized as strings to preserve decimal precision.
    - All timestamps are ISO 8601 UTC.
    - All endpoints are deterministic on a fixed data snapshot per Constitution III.
servers:
  - url: http://localhost:8000
    description: Local dev
paths:
...
  /strategies:
    get:
      summary: List all available strategies
      parameters:
        - { name: timeframe, in: query, required: false, schema: { type: string, enum: [Short-term, Mid-term] } }
        - { name: enabled_only, in: query, required: false, schema: { type: boolean } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  strategies:
                    type: array
                    items: { $ref: '#/components/schemas/Strategy' }
                  data_as_of: { type: string, format: date-time }
                  disclaimer: { type: string }
  /strategies/{slug}:
    get:
      summary: Get one strategy's metadata
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Strategy' }
        '404':
          description: Not found
...
  /strategies/{slug}/run:
    post:
      summary: Run a strategy and return its ScreenResult
      description: |
        Deterministic on a fixed data snapshot — same parameters + filters + as_of_date produce the same id and candidate list.
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ScreenRunRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScreenResult' }
        '400':
          description: Bad parameters
        '404':
          description: Strategy not found
...
components:
  schemas:
...
    Strategy:
      type: object
      required: [slug, name, timeframe, citation]
      properties:
        slug: { type: string }
        name: { type: string }
        timeframe: { type: string, enum: [Short-term, Mid-term] }
        citation: { type: string }
        description: { type: string }
        holding_period_days:
          type: object
          properties:
            min: { type: integer }
            max: { type: integer }
        parameters:
          type: object
          additionalProperties:
            type: object
            properties:
              default: {}
              min: {}
              max: {}
              type: { type: string }
              description: { type: string }
        regime_favorability:
          type: object
          properties:
            "Trending up": { type: string, enum: [Favorable, Neutral, Unfavorable] }
            "Range-bound": { type: string, enum: [Favorable, Neutral, Unfavorable] }
            "Trending down": { type: string, enum: [Favorable, Neutral, Unfavorable] }
        default_exclude_earnings_within_days: { type: integer }
        enabled_by_default: { type: boolean }
        modifications:
          type: array
          description: Each documented modification to the cited core methodology, with its own citation. Powers the "Strategy gates" panel (FR-037).
          items:
            type: object
            properties:
              name: { type: string }
              description: { type: string }
              citation: { type: string }
        backtest_summary:
          type: object
          nullable: true
          description: Headline metrics for the strategy's most recent walk-forward backtest. Null if no backtest has been run.
          properties:
            data_window_start: { type: string, format: date }
            data_window_end: { type: string, format: date }
            total_return: { type: number }
            max_drawdown: { type: number }
            hit_rate: { type: number }
            avg_win: { type: number }
            avg_loss: { type: number }
            turnover: { type: number }
            source_name: { type: string }
            source_as_of: { type: string, format: date-time }
        backtest_window_meets_v1_floor:
          type: boolean
          description: True iff window ≥ 15 years AND covers 2008-2009. Constitution III gate for enabled_by_default.
    ScreenRunRequest:
      type: object
      properties:
        parameters:
          type: object
          additionalProperties: true
          description: Strategy-specific parameter overrides; missing keys use defaults.
        filters:
          type: object
          properties:
            shariah_only: { type: boolean, default: false }
            exclude_earnings_within_days: { type: integer, nullable: true, description: "null → use strategy default" }
        shariah_overrides:
          type: object
          description: |
            Client-supplied user overrides combined with the active external Shariah source(s) per FR-040.
            Server is stateless on user state; lists are stored in the browser and passed per request.
          properties:
            active_sources:
              type: array
              items: { type: string }
              default: [spus_holdings]
            inclusion:
              type: array
              items: { $ref: '#/components/schemas/UserShariahOverride' }
            exclusion:
              type: array
              items: { $ref: '#/components/schemas/UserShariahOverride' }
        as_of_date:
          type: string
          format: date
          description: Optional override; defaults to the latest available data snapshot.
    ScreenResult:
      type: object
      required: [id, strategy_slug, as_of_date, candidates, data_as_of, disclaimer]
      properties:
        id: { type: string, description: Deterministic hash of (strategy_slug, as_of_date, parameters, filters). }
        strategy_slug: { type: string }
        as_of_date: { type: string, format: date }
        parameters_snapshot:
          type: object
          additionalProperties: true
        filters_snapshot:
          type: object
          properties:
            shariah_only: { type: boolean }
            exclude_earnings_within_days: { type: integer }
            shariah_active_sources:
              type: array
              items: { type: string }
            shariah_user_inclusion:
              type: array
              items: { $ref: '#/components/schemas/UserShariahOverride' }
            shariah_user_exclusion:
              type: array
              items: { $ref: '#/components/schemas/UserShariahOverride' }
        candidate_count: { type: integer }
        candidates:
          type: array
          items: { $ref: '#/components/schemas/Candidate' }
        computed_at: { type: string, format: date-time }
        data_as_of: { type: string, format: date-time }
        disclaimer: { type: string }
        stale_sources:
          type: array
          items: { type: string }
          description: Source names whose data is older than their refresh interval.
        data_notes:
          type: array
          items: { type: string }
          description: >-
            Honest data-completeness notes: strategy gates skipped because required
            data was missing (fundamentals coverage, volume history, sector breadth)
            and price-bar staleness warnings. Empty when every declared gate ran on
            fresh data.
    Candidate:
      type: object
      required: [ticker, name, sector, current_price, entry, stop_loss, take_profit, rank, reason]
      properties:
        ticker: { type: string }
        name: { type: string }
        sector: { type: string }
        current_price: { type: string, description: USD decimal as string. }
        entry: { type: string }
        stop_loss: { type: string }
        tighter_stop_loss:
          type: string
          nullable: true
          description: Alternative tighter stop from the structure (20-day swing-low) method, shown beside the primary stop for comparison. Null when unavailable.
        take_profit: { type: string }
        rank: { type: integer, minimum: 1 }
        score: { type: number }
        reason: { type: string, description: One-line "why it matched". }
        gate_results:
          type: array
          description: Per-gate breakdown for this candidate — every gate that ran (pass, with the actual value) and every gate skipped for missing data. A listed candidate passed all applied gates.
          items:
            type: object
            properties:
              gate: { type: string }
              status: { type: string, enum: [pass, fail, skipped] }
              detail: { type: string }
        shariah_compliant:
          type: boolean
          nullable: true
          description: Null when filter is OFF; true when filter is ON (false candidates are excluded).
        shariah_source_kind:
          type: string
          nullable: true
          enum: [external, user, excluded_by_user, not_listed]
        shariah_external_source_name:
          type: string
          nullable: true
        shariah_source_as_of:
          type: string
          format: date-time
          nullable: true
        shariah_source_url:
          type: string
          nullable: true
        shariah_user_note:
          type: string
          nullable: true
        shariah_is_stale:
          type: boolean
          nullable: true
        next_earnings_date:
          type: string
          format: date
          nullable: true
        days_to_earnings:
          type: integer
          nullable: true
        recent_8k_count_30d: { type: integer }
        events_source_as_of:
          type: string
          format: date-time
          nullable: true
...
    UserShariahOverride:
      type: object
      required: [ticker, direction]
      properties:
        ticker: { type: string }
        direction:
          type: string
          enum: [include, exclude]
        added_at: { type: string, format: date-time, nullable: true }
        note:
          type: string
          nullable: true
          description: User's free-text rationale (e.g., link to a Shariah-board review they trust).
...